requirements.txt
flask
flask_socketio
eventlet
imagehash
opencv-python
Pillow
icrawler
Werkzeug
